+++ trunk/jdem846/src/us/wthr/jdem846/color/GrayTint.java
@@ -57,8 +57,8 @@ public class GrayTint implements ModelColoring
 	}
 
 	@Override
-	public void getGradientColor(float elevation, float minElevation,
-			float maxElevation, int[] color) 
+	public void getGradientColor(double elevation, double minElevation,
+			double maxElevation, int[] color) 
 	{
 		defaultColor.toList(color);
 		//return defaultColor.getCopy();
