+++ trunk/jdem846/src/us/wthr/jdem846/input/bil/BilInt16.java
@@ -39,10 +39,10 @@ public class BilInt16 extends DataSource
 	
 	private int bits = 16;
 	private int totalRowBytes;
-	private float maxDifference;
-	private float maxElevation;
-    private float minElevation;
-    private float resolution;
+	private double maxDifference;
+	private double maxElevation;
+    private double minElevation;
+    private double resolution;
     private int maxRow;
     private int maxCol;
     private String filePath;
@@ -133,7 +133,7 @@ public class BilInt16 extends DataSource
 		cache.unload();
 	}
 
-	private float __get(int row, int col)
+	private double __get(int row, int col)
 	{
 		if (col < 0 || col > header.getColumns() || row < 0 || row > header.getRows())
 			return DemConstants.ELEV_NO_DATA;
@@ -150,12 +150,12 @@ public class BilInt16 extends DataSource
 			log.warn("Column to high!!! -- " + row + "/" + col + " -- " + header.getColumns() + " -- " + col);
 		}
 		
-		float elevation = cache.get(col);
+		double elevation = cache.get(col);
 		return elevation;
 	}
 	
 	
-	public float getElevation(int col)
+	public double getElevation(int col)
 	{
 		if (col >= header.getColumns())
 			return DemConstants.ELEV_NO_DATA;
@@ -163,14 +163,14 @@ public class BilInt16 extends DataSource
 		return this.getElevation(cachedRow, col);
 	}
 
-	public float getElevation(int row, int column)
+	public double getElevation(int row, int column)
 	{
 		if (column >= header.getColumns() || row > header.getRows())
 			return DemConstants.ELEV_NO_DATA;
 		
 		
 		
-		float elevation = this.__get(row, column);
+		double elevation = this.__get(row, column);
 		if (elevation == header.getNoData())
 			elevation = DemConstants.ELEV_NO_DATA;
 
@@ -207,42 +207,42 @@ public class BilInt16 extends DataSource
 		this.totalRowBytes = totalRowBytes;
 	}
 
-	public float getMaxDifference()
+	public double getMaxDifference()
 	{
 		return maxDifference;
 	}
 
-	public void setMaxDifference(float maxDifference)
+	public void setMaxDifference(double maxDifference)
 	{
 		this.maxDifference = maxDifference;
 	}
 
-	public float getMaxElevation() 
+	public double getMaxElevation() 
 	{
 		return maxElevation;
 	}
 
-	public void setMaxElevation(float maxElevation)
+	public void setMaxElevation(double maxElevation)
 	{
 		this.maxElevation = maxElevation;
 	}
 
-	public float getMinElevation() 
+	public double getMinElevation() 
 	{
 		return minElevation;
 	}
 
-	public void setMinElevation(float minElevation) 
+	public void setMinElevation(double minElevation) 
 	{
 		this.minElevation = minElevation;
 	}
 
-	public float getResolution()
+	public double getResolution()
 	{
 		return resolution;
 	}
 
-	public void setResolution(float resolution) 
+	public void setResolution(double resolution) 
 	{
 		this.resolution = resolution;
 	}
